requirements.txt
# SEC 10-K RAG — Python stack (CLI + ingestion + OpenAI embeddings + Chroma)

chromadb>=0.4.24,<0.6
openai>=1.40.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

pytest>=7.4.0

# Embeddings: OpenAI text-embedding-3-small (see src/Embed/embed.py).
